requirements.txt
Flask==3.0.3
gspread==6.1.2
google-auth==2.34.0
Pillow==10.4.0
qrcode==7.4.2
gunicorn==22.0.0
python-dotenv==1.0.1
oauth2client==4.1.3
imagefont==1.0.3